RF 05 e 06 prontas p/ teste

+++ CleanFinance/src/main/java/br/com/CleanFinance/Cartao/CartaoRepository.java
@@ -2,5 +2,5 @@ package br.com.CleanFinance.Cartao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CartaoRepository extends JpaRepository<Cartao,Integer> {
+public interface CartaoRepository extends JpaRepository<Cartao,Long> {
 }
